--- app/src/apis/watch-list/index.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const ctrl = require("./watch-list.ctrl");
-
-router.get("/:studentId", ctrl.process.findAllByStudentId);
-router.delete("/:studentId", ctrl.process.delete);
-router.post("/:studentId", ctrl.process.update);
-
-module.exports = router;
